Fix issues with CaDDN (#542)
* Download DeepLabV3 if not available yet

* Fix dtype issue

Fix RuntimeError: expected backend CPU and dtype Float but got backend CPU and dtype Long

+++ pcdet/models/backbones_3d/vfe/image_vfe_modules/f2v/frustum_grid_generator.py
@@ -17,7 +17,7 @@ class FrustumGridGenerator(nn.Module):
         """
         super().__init__()
         self.dtype = torch.float32
-        self.grid_size = torch.as_tensor(grid_size)
+        self.grid_size = torch.as_tensor(grid_size, dtype=self.dtype)
         self.pc_range = pc_range
         self.out_of_bounds_val = -2
         self.disc_cfg = disc_cfg
